// src/lib/types.ts
export interface ReactFiber {
  tag: number;
  type: unknown;
  child: ReactFiber | null;
  sibling: ReactFiber | null;
  return: ReactFiber | null;
  stateNode: Node | null;
  pendingProps: Record<string, unknown> | null;
  memoizedProps: Record<string, unknown> | null;
  _debugOwner?: ReactFiber;
  _debugSource?: { fileName: string; lineNumber: number };
}
// ...
export interface SerializedFiberNode {
  h2dId: string | undefined;
  name: string | undefined;
  fiberTag: number | null;
  props: Record<string, unknown> | undefined;
  children: SerializedFiberNode[];
}

export interface TruncatedPropValue {
  truncated: true;
  value: string;
  originalLength: number;
}

export interface PropRefValue {
  ref: string;
}